Update tasks 1 to 4 to use React 16
This commit rewrites all code solutions to modern practices to support
the update to React 16.

To simplify, the code is done directly in the index.html file as well,
so it's easier to hack away in your browser.

Finally, the docs are updated with up to date links and tasks.